+++ README.md
@@ -1,0 +1,24 @@
+# InvUnload
+
+
+Automatically puts your stuff into the right chests
+
+You don't have time to open every chest in your storage room to tidy up your inventory?
+Fear no more! You have no more excuses for having chests full of random garbage!
+
+<p align="center">
+  <img src="https://www.spigotmc.org/attachments/if_hand_truck_267484-png.367171/"/>
+</p>
+
+InvUnload does two brilliant things: When you enter /unload, it checks if there are chests nearby.
+
+For each chest, the player's inventory will be searched for matching items. If there are any, they will be put into the chest.
+## Build
+
+To build the project:
+
+1. Install Maven
+
+2. Put [ChestSort.jar](https://www.spigotmc.org/resources/1-11-1-13-chestsort-api.59773/download?version=258008) into the project at `lib/ChestSort.jar`.
+
+3. Run `mvn clean package`
