+++ src/main/java/de/jeffclan/InvUnload/Utils.java
@@ -1,0 +1,15 @@
+package de.jeffclan.InvUnload;
+
+public class Utils {
+	public static boolean isInteger(String s) {
+	    try { 
+	        Integer.parseInt(s); 
+	    } catch(NumberFormatException e) { 
+	        return false; 
+	    } catch(NullPointerException e) {
+	        return false;
+	    }
+
+	    return true;
+	}
+}
