+++ src/main/resources/plugin.yml
@@ -1,0 +1,23 @@
+main: de.jeffclan.InvUnload.InvUnload
+name: InvUnload
+version: 2.0.2
+api-version: 1.13
+description: Unloads inventory in nearby matching chests
+author: mfnalex
+prefix: InvUnload
+database: false
+softdepend: [WorldGuard]
+commands:
+  unload:
+    description: Unloads inventory in nearby matching chests
+    usage: /<command> [radius]
+    permission: invunload.unload
+  dump:
+    description: Unloads inventory in nearby chests
+    usage: /<command> [radius]
+    permission: invunload.dump
+permissions:
+  invunload.unload:
+    description: Allows usage of command /unload
+  invunload:dump:
+    description: Allows usage of command /dump
